requirements.txt
requests>=2.28
Pillow>=9.0
piexif>=1.1
mutagen>=1.45
Flask>=2.3
Flask-Login>=0.6
tqdm>=4.60
boto3>=1.26
python-dotenv>=1.0
gunicorn>=21.0
